@@ -19,22 +19,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.5",
-    "@angular/common": "^4.4.5",
-    "@angular/compiler": "^4.4.5",
-    "@angular/core": "^4.4.5",
-    "@angular/forms": "^4.4.5",
-    "@angular/http": "^4.4.5",
-    "@angular/platform-browser": "^4.4.5",
-    "@angular/platform-browser-dynamic": "^4.4.5",
-    "@angular/router": "^4.4.5",
-    "@types/lodash": "^4.14.88",
-    "commander": "^2.12.2",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^5.2.9",
+    "@angular/common": "^5.2.9",
+    "@angular/compiler": "^5.2.9",
+    "@angular/core": "^5.2.9",
+    "@angular/forms": "^5.2.9",
+    "@angular/http": "^5.2.9",
+    "@angular/platform-browser": "^5.2.9",
+    "@angular/platform-browser-dynamic": "^5.2.9",
+    "@angular/router": "^5.2.9",
+    "@types/lodash": "^4.14.106",
+    "commander": "^2.15.1",
+    "core-js": "^2.5.5",
     "html2pdf.js": "^0.8.2",
-    "js-base64": "^2.4.0",
-    "js-yaml": "^3.10.0",
-    "lodash-es": "^4.17.4",
+    "js-base64": "^2.4.3",
+    "js-yaml": "^3.11.0",
+    "lodash-es": "^4.17.8",
     "loglevel": "^1.6.1",
     "mathsass": "^0.10.1",
     "mime": "^1.6.0",
@@ -43,22 +43,23 @@
     "ng2-nouislider": "1.7.4",
     "nouislider": "^10.1.0",
     "regenerator-runtime": "^0.11.1",
-    "request": "^2.83.0",
-    "rxjs": "^5.5.5",
-    "statuses": "^1.4.0",
+    "request": "^2.85.0",
+    "rxjs": "^5.5.9",
+    "statuses": "^1.5.0",
     "tmp": "0.0.33",
-    "url-parse": "^1.2.0",
+    "url-parse": "^1.3.0",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.14"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.0",
-    "@angular/compiler-cli": "^4.4.5",
-    "@angular/language-service": "^4.4.5",
-    "@types/jasmine": "^2.8.2",
+    "@angular/cli": "^1.7.4",
+    "@angular/compiler": "^5.2.9",
+    "@angular/compiler-cli": "^5.2.9",
+    "@angular/language-service": "^5.2.9",
+    "@types/jasmine": "^2.8.6",
     "@types/jasminewd2": "~2.0.2",
     "@types/lodash-es": "^4.17.0",
-    "@types/node": "~6.0.60",
+    "@types/node": "^6.0.105",
     "@types/url-parse": "^1.1.0",
     "codelyzer": "~3.1.1",
     "engine-strict": "^1.0.0",
@@ -67,14 +68,15 @@
     "karma": "~1.7.0",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "markdown-toc": "^1.1.0",
-    "node-sass": "^4.6.0",
+    "node-sass": "^4.8.3",
     "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3",
-    "webpack-bundle-analyzer": "^2.9.0"
-  }
+    "tslint": "~5.9.1",
+    "typescript": "^2.6.2",
+    "webpack-bundle-analyzer": "^2.11.1"
+  },
+  "bundleDependencies": []
 }
