@@ -17,12 +17,19 @@ import {MeasurePageModule} from './measure-page/measure-page.module';
 import {BoilerModule} from './boiler/boiler.module';
 import {GrantsModule} from './grants/grants.module';
 import {EnergyEfficiencyModule} from './energy-efficiency/energy-efficiency.module';
-import {AdminPageModule} from "./admin-page/admin-page.module";
+import {AdminPageModule} from './admin-page/admin-page.module';
+import {YourHomeModule} from './your-home/your-home.module';
 import {MeesModule} from './mees/mees.module';
+import {ECOSuppliersPageModule} from "./eco-suppliers-page/eco-suppliers-page.module";
+import {PageNotFoundComponent} from './page-not-found/page-not-found.component';
+import {PageNotCreatedComponent} from "./page-not-created/page-not-created.component";
+import {SimpleSavingsModule} from './simple-savings/simple-savings.module';
 
 @NgModule({
     declarations: [
         AppComponent,
+        PageNotFoundComponent,
+        PageNotCreatedComponent
     ],
     imports: [
         BrowserModule,
@@ -37,12 +44,15 @@ import {MeesModule} from './mees/mees.module';
         GrantsModule.forRoot(),
         PageModule,
         MeasurePageModule,
+        ECOSuppliersPageModule,
         AdminPageModule,
         LandingPageModule,
         LayoutComponentsModule.forRoot(),
         BoilerModule.forRoot(),
+        InlineSVGModule,
+        YourHomeModule,
         MeesModule.forRoot(),
-        InlineSVGModule
+        SimpleSavingsModule,
     ],
     bootstrap: [AppComponent],
 })
