@@ -18,9 +18,9 @@ export class BoilerLandingPageComponent {
 
     onAddressSelected(lmkKey: string) {
         if (lmkKey) {
-            this.router.navigate(['/js/boiler/epc-replace', lmkKey]);
+            this.router.navigate(['/boiler/epc-replace', lmkKey]);
         } else {
-            this.router.navigate(['/js/boiler/epc-replace/unknown']);
+            this.router.navigate(['/boiler/epc-replace/unknown']);
         }
     }
 
