@@ -15,7 +15,7 @@ export class BoilerResultsPageRouteGuard implements CanActivate {
         } else {
             // TODO:BEISDEAS-201 display a user-visible error here
             console.error("Cannot show results; questions are incomplete");
-            this.router.navigate(['/js/boiler/questionnaire']);
+            this.router.navigate(['/boiler/questionnaire']);
             return false;
         }
     }
