@@ -42,6 +42,7 @@ export function getTagsForMeasure(measureContent: MeasureContent): EnergyEfficie
     if (measureContent.acf.tags) {
         tags = measureContent.acf.tags
             .map(tagName => RECOMMENDATION_TAGS_BY_JSON_NAME[tagName])
+            .filter(x => x)
             .reduce((result, value) => {
                 result |= value;
                 return result;
@@ -55,5 +56,3 @@ const RECOMMENDATION_TAGS_BY_JSON_NAME = {
     tag_small_spend: EnergyEfficiencyRecommendationTag.SmallSpend,
     tag_longer_term: EnergyEfficiencyRecommendationTag.LongerTerm
 };
-
-export type RecommendationTagJsonName = keyof typeof RECOMMENDATION_TAGS_BY_JSON_NAME;
