@@ -1,5 +1,4 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
-import {InlineSVGModule} from 'ng-inline-svg';
 import {HomePageComponent} from './home-page.component';
 import {LatestNewsCardComponent} from '../shared/latest-news-card/latest-news-card.component';
 import {ResponseData} from '../shared/response-data/response-data';
@@ -9,10 +8,14 @@ import {UserJourneyType} from '../shared/response-data/user-journey-type';
 import {Location} from '@angular/common';
 import {NavigationBarComponent} from '../layout-components/navigation-bar/navigation-bar.component';
 import {QuestionnaireService} from '../questionnaire/questionnaire.service';
+import {FormsModule} from '@angular/forms';
+import {InlineSVGModule} from 'ng-inline-svg';
 import {Observable} from 'rxjs/Observable';
 import {WordpressPagesService} from '../shared/wordpress-pages-service/wordpress-pages.service';
 import {LatestNewsSectionComponent} from '../shared/latest-news-section/latest-news-section.component';
+import {SearchBarComponent} from "../layout-components/search-bar/search-bar.component";
 import {NavBarSuboptionComponent} from "../layout-components/navigation-bar/nav-bar-suboption/nav-bar-suboption.component";
+import {GoogleAnalyticsService} from "../shared/analytics/google-analytics.service";
 
 describe('HomePageComponent', () => {
     let component: HomePageComponent;
@@ -27,21 +30,18 @@ describe('HomePageComponent', () => {
     beforeEach(async(() => {
         responseDataStub = new ResponseData();
         TestBed.configureTestingModule({
-            declarations: [HomePageComponent, LatestNewsSectionComponent, LatestNewsCardComponent, NavigationBarComponent,
-                NavBarSuboptionComponent],
-            imports: [
-                RouterTestingModule.withRoutes([
-                    {path: 'js/energy-efficiency/questionnaire/home-basics', component: DummyComponent},
-                    {path: 'js/energy-efficiency/results', component: DummyComponent},
-                    {path: 'js/grants', component: DummyComponent},
-                    {path: 'js/boiler', component: DummyComponent},
-                    {path: 'js/reduce-bills', component: DummyComponent},
-                    {path: 'js/warmer-home', component: DummyComponent},
-                    {path: 'js/home-improvements', component: DummyComponent},
-                    {path: 'js/greener-home', component: DummyComponent}
-                ]),
-                InlineSVGModule,
-            ],
+            declarations: [HomePageComponent, LatestNewsSectionComponent,
+                LatestNewsCardComponent, NavBarSuboptionComponent, NavigationBarComponent, SearchBarComponent],
+            imports: [RouterTestingModule.withRoutes([
+                {path: 'energy-efficiency/questionnaire/home-basics', component: DummyComponent},
+                {path: 'energy-efficiency/results', component: DummyComponent},
+                {path: 'grants', component: DummyComponent},
+                {path: 'boiler', component: DummyComponent},
+                {path: 'energy-efficiency/reduce-bills', component: DummyComponent},
+                {path: 'energy-efficiency/warmer-home', component: DummyComponent},
+                {path: 'energy-efficiency/home-improvements', component: DummyComponent},
+                {path: 'energy-efficiency/greener-home', component: DummyComponent},
+            ]), FormsModule, InlineSVGModule],
             providers: [
                 {provide: ResponseData, useValue: responseDataStub},
                 {
@@ -51,7 +51,9 @@ describe('HomePageComponent', () => {
                 {
                     provide: WordpressPagesService,
                     useValue: {getLatestPages: () => Observable.of([])}
-                }]
+                },
+                GoogleAnalyticsService,
+            ]
         })
             .compileComponents();
     }));
@@ -67,19 +69,19 @@ describe('HomePageComponent', () => {
     });
 
     it('should set the user journey type and move to the reduce bills landing page when the link is clicked', async () => {
-        checkLink('.landing-page-link.reduce-bills', '/js/reduce-bills');
+        checkLink('.landing-page-link.reduce-bills', '/energy-efficiency/reduce-bills');
     });
 
     it('should set the user journey type and move to the warmer home landing page when the link is clicked', async () => {
-        checkLink('.landing-page-link.warmer-home', '/js/warmer-home');
+        checkLink('.landing-page-link.warmer-home', '/energy-efficiency/warmer-home');
     });
 
     it('should set the user journey type and move to the home improvements landing page when the link is clicked', async () => {
-        checkLink('.landing-page-link.home-improvements', '/js/home-improvements');
+        checkLink('.landing-page-link.home-improvements', '/energy-efficiency/home-improvements');
     });
 
     it('should set the user journey type and move to the greener home landing page when the link is clicked', async () => {
-        checkLink('.landing-page-link.greener-home', '/js/greener-home');
+        checkLink('.landing-page-link.greener-home', '/energy-efficiency/greener-home');
     });
 
     it('should set the user journey type and move to the questionnaire when the appropriate button is clicked, if questionnaire incomplete',
@@ -87,7 +89,7 @@ describe('HomePageComponent', () => {
         // given
         questionnaireComplete = false;
 
-        checkLink('.calculator', '/js/energy-efficiency/questionnaire/home-basics', UserJourneyType.Calculator);
+        checkLink('.calculator', '/energy-efficiency/questionnaire/home-basics', UserJourneyType.Calculator);
     });
 
     it('should set the user journey type and move to the results page when the appropriate button is clicked, if questionnaire complete',
@@ -95,15 +97,15 @@ describe('HomePageComponent', () => {
         // given
         questionnaireComplete = true;
 
-        checkLink('.calculator', '/js/energy-efficiency/results', UserJourneyType.Calculator);
+        checkLink('.calculator', '/energy-efficiency/results', UserJourneyType.Calculator);
     });
 
     it('should set the user journey type and move to the grants page when the appropriate button is clicked', async () => {
-        checkLink('.grants', '/js/grants', UserJourneyType.Grants);
+        checkLink('.grants', '/grants', UserJourneyType.Grants);
     });
 
     it('should set the user journey type and move to the boilers page when the appropriate button is clicked', async () => {
-        checkLink('.boiler', '/js/boiler', UserJourneyType.Boiler);
+        checkLink('.boiler', '/boiler', UserJourneyType.Boiler);
     });
 
     const checkLink = (selector: string, path: string, journeyType?: UserJourneyType) => {
