@@ -3,6 +3,7 @@ import {UserJourneyType} from '../shared/response-data/user-journey-type';
 import {ResponseData} from '../shared/response-data/response-data';
 import {Router} from '@angular/router';
 import {QuestionnaireService} from '../questionnaire/questionnaire.service';
+import {GoogleAnalyticsService} from "../shared/analytics/google-analytics.service";
 
 @Component({
     selector: 'app-home-page',
@@ -13,24 +14,32 @@ export class HomePageComponent {
 
     constructor(private responseData: ResponseData,
                 private questionnaireService: QuestionnaireService,
+                private googleAnalyticsService: GoogleAnalyticsService,
                 private router: Router) {
     }
 
     onEnergyCalculatorButtonClick() {
+        this.sendEventToAnalytics('calculator_clicked');
         this.responseData.userJourneyType = UserJourneyType.Calculator;
         const route = this.questionnaireService.isComplete('home-basics')
-            ? '/js/energy-efficiency/results'
-            : '/js/energy-efficiency/questionnaire/home-basics';
+            ? '/energy-efficiency/results'
+            : '/energy-efficiency/questionnaire/home-basics';
         this.router.navigate([route]);
     }
 
     onBoilerButtonClick() {
+        this.sendEventToAnalytics('boilers_clicked');
         this.responseData.userJourneyType = UserJourneyType.Boiler;
-        this.router.navigate(['/js/boiler']);
+        this.router.navigate(['/boiler']);
     }
 
     onGrantsButtonClick() {
+        this.sendEventToAnalytics('grants_clicked');
         this.responseData.userJourneyType = UserJourneyType.Grants;
-        this.router.navigate(['/js/grants']);
+        this.router.navigate(['/grants']);
+    }
+
+    sendEventToAnalytics(eventName: string) {
+        this.googleAnalyticsService.sendEvent(eventName, 'home-page');
     }
 }
