@@ -3,7 +3,7 @@ import {QuestionnairePage} from "../questionnaire-page";
 
 export class BehaviourQuestionnairePage extends QuestionnairePage{
     navigateTo() {
-        return browser.get('/js/energy-efficiency/questionnaire/behaviour');
+        return browser.get('/energy-efficiency/questionnaire/behaviour');
     }
 
     selectShowerType() {
