@@ -6,7 +6,7 @@ export class HomeBasicsQuestionnairePage extends QuestionnairePage {
     private miniEpcEnabled: boolean;
 
     navigateTo() {
-        return browser.get('/js/energy-efficiency/questionnaire/home-basics');
+        return browser.get('/energy-efficiency/questionnaire/home-basics');
     }
 
     enterPostcode(postcode: string) {
